Make delete requests require an Etag

+++ src/IIIFPresentation/API/Features/Manifest/Requests/DeleteManifest.cs
@@ -1,4 +1,6 @@
-﻿using API.Features.Storage.Helpers;
+﻿using API.Features.Common.Helpers;
+using API.Features.Storage.Helpers;
+using API.Infrastructure.Helpers;
 using AWS.Helpers;
 using Core;
 using MediatR;
@@ -8,20 +10,21 @@ using Repository;
 
 namespace API.Features.Manifest.Requests;
 
-public class DeleteManifest(int customerId, string manifestId)
-    : IRequest<ResultMessage<DeleteResult, DeleteManifestType>>
+public class DeleteManifest(int customerId, string manifestId, string etag)
+    : IRequest<ResultMessage<DeleteResult, DeleteResourceType>>
 {
     public int CustomerId { get; } = customerId;
     public string ManifestId { get; } = manifestId;
+    public string? Etag { get; } = etag;
 }
 
 public class DeleteManifestHandler(
     PresentationContext dbContext,
     IIIIFS3Service iiifS3,
     ILogger<DeleteManifestHandler> logger)
-    : IRequestHandler<DeleteManifest, ResultMessage<DeleteResult, DeleteManifestType>>
+    : IRequestHandler<DeleteManifest, ResultMessage<DeleteResult, DeleteResourceType>>
 {
-    public async Task<ResultMessage<DeleteResult, DeleteManifestType>> Handle(DeleteManifest request,
+    public async Task<ResultMessage<DeleteResult, DeleteResourceType>> Handle(DeleteManifest request,
         CancellationToken cancellationToken)
     {
         logger.LogDebug("Deleting manifest {ManifestId} for customer {CustomerId}", request.ManifestId,
@@ -32,6 +35,8 @@ public class DeleteManifestHandler(
                 withCanvasPaintings: false, cancellationToken: cancellationToken);
         
         if (manifest is null) return new(DeleteResult.NotFound);
+        
+        if (!EtagComparer.IsMatch(manifest.Etag, request.Etag)) return DeleteErrorHelper.EtagNotMatching();
 
         dbContext.Manifests.Remove(manifest);
 
@@ -46,7 +51,7 @@ public class DeleteManifestHandler(
             logger.LogError(ex, "Error attempting to delete manifest {ManifestId} for customer {CustomerId}",
                 request.ManifestId, request.CustomerId);
             return new(DeleteResult.Error,
-                DeleteManifestType.Unknown, "Error deleting manifest");
+                DeleteResourceType.Unknown, "Error deleting manifest");
         }
         
         return new(DeleteResult.Deleted);
